Move tests out of package
