+++ anndata/tests/__init__.py
@@ -1,0 +1,3 @@
+import warnings
+
+warnings.warn("Please import from anndata.test_utils instead", FutureWarning)
