+++ tests/test_get_vector.py
